+++ elide-core/src/test/java/com/yahoo/elide/core/RequestScopeTest.java
@@ -5,13 +5,18 @@
  */
 package com.yahoo.elide.core;
 
+import com.yahoo.elide.ElideSettingsBuilder;
+import com.yahoo.elide.annotation.Include;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 import java.lang.reflect.Method;
 import java.util.Collections;
+import java.util.HashMap;
 
 public class RequestScopeTest {
     @Test
@@ -29,5 +34,40 @@ public class RequestScopeTest {
         Assert.assertEquals(filtered.size(), 2);
         Assert.assertTrue(filtered.containsKey("filter"));
         Assert.assertTrue(filtered.containsKey("filter[xyz]"));
+    }
+
+    @Test
+    public void testNewObjectsForInheritedTypes() throws Exception {
+        // NOTE: This tests that inherited types are properly accounted for during a patch extension request.
+        //       otherwise it is possible to create an inherited type and when performing introspection on a
+        //       superclass we will miss that the object is newly created and then we'll attempt to query
+        //       the datastore.
+        EntityDictionary dictionary = new EntityDictionary(new HashMap<>());
+
+        dictionary.bindEntity(MyBaseClass.class);
+        dictionary.bindEntity(MyInheritedClass.class);
+
+        RequestScope requestScope = new RequestScope("/", null, null, null, null,
+                new ElideSettingsBuilder(null)
+                        .withEntityDictionary(dictionary)
+                        .build());
+
+        String myId = "myId";
+        // Test that a new inherited class is counted for base type
+        requestScope.setUUIDForObject(dictionary.getJsonAliasFor(MyInheritedClass.class), myId, new MyInheritedClass());
+        Assert.assertNotNull(requestScope.getObjectById(dictionary.getJsonAliasFor(MyBaseClass.class), myId));
+    }
+
+    @Entity
+    @Include
+    public static class MyBaseClass {
+        @Id
+        public long id;
+    }
+
+    @Entity
+    @Include
+    public static class MyInheritedClass extends MyBaseClass {
+        public String myField;
     }
 }